refactor(cli): extract handoff complexity owner
